backend/requirements.txt
Flask==2.2.2
google-cloud-storage==2.5.0
google-cloud-firestore==2.7.2
python-dotenv==0.21.0
Flask-Cors==3.0.10
gunicorn==20.1.0